+++ client/Packages/com.beamable.server/Runtime/MicroserviceClient.cs
@@ -56,7 +56,7 @@ namespace Beamable.Server
 		protected async Promise<T> Request<T>(string serviceName, string endpoint, string[] serializedFields)
 		{
 			var requester = _requester ?? await API.Instance.Map(b => b.Requester);
-			return await MicroserviceClientHelper.Request<T>(requester, serviceName, endpoint, serializedFields);
+			return await MicroserviceClientHelper.Request<T>(Provider, requester, serviceName, endpoint, serializedFields);
 		}
 
 		protected string SerializeArgument<T>(T arg) => MicroserviceClientHelper.SerializeArgument(arg);
@@ -290,12 +290,30 @@ namespace Beamable.Server
 		public static string CreateUrl(string cid, string pid, string serviceName, string endpoint)
 		{
 			var prefix = _prefix ?? (_prefix = MicroserviceIndividualization.GetServicePrefix(serviceName));
+			return CreateUrl(cid, pid, serviceName, endpoint, prefix);
+		}
+		
+		public static string CreateUrl(string cid, string pid, string serviceName, string endpoint, string prefix)
+		{
 			var path = $"{prefix}micro_{serviceName}/{endpoint}";
 			var url = $"/basic/{cid}.{pid}.{path}";
 			return url;
 		}
 
-		public static async Promise<T> Request<T>(IBeamableRequester requester, string serviceName, string endpoint, string[] serializedFields)
+		public static Dictionary<string, Promise<string>> serviceNameToPrefixPromise =
+			new Dictionary<string, Promise<string>>();
+
+		[Obsolete("Use the variant that accepts a dependency provider.")]
+		public static Promise<T> Request<T>(IBeamableRequester requester,
+		                                          string serviceName,
+		                                          string endpoint,
+		                                          string[] serializedFields)
+		{
+			var ctx = BeamContext.Default;
+			return Request<T>(ctx.ServiceProvider, requester, serviceName, endpoint, serializedFields);
+		}
+
+		public static async Promise<T> Request<T>(IDependencyProvider provider, IBeamableRequester requester, string serviceName, string endpoint, string[] serializedFields)
 		{
 			var argArray = "[ " + string.Join(",", serializedFields) + " ]";
 
@@ -312,9 +330,23 @@ namespace Beamable.Server
 				return result;
 			}
 
-			// var b = await API.Instance;
-			// var requester = b.Requester;
-			var url = CreateUrl(requester.AccessToken.Cid, requester.AccessToken.Pid, serviceName, endpoint);
+			if (!serviceNameToPrefixPromise.TryGetValue(serviceName, out var prefixPromise))
+			{
+				// need to resolve the IPrefixService
+				if (provider.CanBuildService<IMicroservicePrefixService>())
+				{
+					var prefixService = provider.GetService<IMicroservicePrefixService>();
+					prefixPromise = serviceNameToPrefixPromise[serviceName] = prefixService.GetPrefix(serviceName);
+				}
+				else
+				{
+					prefixPromise = Promise<string>.Successful("");
+				}
+			}
+
+			var prefix = await prefixPromise;
+			var url = CreateUrl(requester.AccessToken.Cid, requester.AccessToken.Pid, serviceName, endpoint, prefix);
+			
 			var req = new RequestObject
 			{
 				payload = argArray